% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{as_messages}
\alias{as_messages}
\title{Convert object into a messages list}
\usage{
as_messages(messages, ...)
}
\arguments{
\item{messages}{object to convert to messages}

\item{...}{These dots are for future extensions and must be empty.}
}
\description{
Convert object into a messages list
}
\examples{
as_messages("hello")
as_messages(list("hello"))
as_messages(list(assistant = "hello", user = "hello"))

}
