% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat.R, R/stream.R
\name{chat}
\alias{chat}
\alias{stream}
\title{Chat with the Mistral api}
\usage{
chat(..., model = "mistral-tiny", error_call = current_env())

stream(..., model = "mistral-tiny", error_call = current_env())
}
\arguments{
\item{...}{messages, see \code{\link[=as_messages]{as_messages()}}.}

\item{model}{which model to use. See \code{\link[=models]{models()}} for more information about which models are available.}

\item{error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\value{
A tibble with columns \code{role} and \code{content} with class \code{chat_tibble}
}
\description{
Chat with the Mistral api

stream
}
\examples{

\dontrun{
  res <- chat("What are the top 5 R packages ?")
  res

  # use the result from a previous chat() to continue the
  # conversation
  chat(res, "Why do people love them so much ?")
}

}
