% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{models}
\alias{models}
\title{Retrieve all models available in the Mistral API}
\usage{
models(error_call = caller_env(), dry_run = FALSE)
}
\arguments{
\item{error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}

\item{dry_run}{if TRUE the request is not performed}
}
\value{
A character vector with the models available in the Mistral API
}
\description{
Retrieve all models available in the Mistral API
}
\examples{
models(dry_run = TRUE)

}
