% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{as_messages}
\alias{as_messages}
\title{Convert object into a messages list}
\usage{
as_messages(..., error_call = current_env())
}
\arguments{
\item{...}{objects to convert to messages. Each element can be:
a string or a result from \code{\link[=chat]{chat()}}.}

\item{error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
Convert object into a messages list
}
\examples{
# unnamed string means user
as_messages("hello")

# explicit names
as_messages(assistant = "hello", user = "hello")

\dontrun{
  res <- chat("hello")

  # add result from previous chat()
  as_messages(res, "hello")
}
}
