% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/messages.R
\name{as_messages}
\alias{as_messages}
\title{Convert object into a messages list}
\usage{
as_messages(messages, ..., error_call = current_env())
}
\arguments{
\item{messages}{object to convert to messages}

\item{...}{ignored}

\item{error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\description{
Convert object into a messages list
}
\examples{
as_messages("hello")
as_messages(list("hello"))
as_messages(list(assistant = "hello", user = "hello"))

}
