% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat.R
\name{chat}
\alias{chat}
\title{Chat with the Mistral api}
\usage{
chat(text = "What are the top 5 R packages ?", model = "mistral-tiny")
}
\arguments{
\item{text}{some text}

\item{which}{model to use. See \code{\link[=models]{models()}} for more information about which models are available}
}
\value{
Result text from Mistral
}
\description{
Chat with the Mistral api
}
\examples{
chat("Top 5 R packages")

}
