% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chat.R, R/stream.R
\name{chat}
\alias{chat}
\alias{stream}
\title{Chat with the Mistral api}
\usage{
chat(
  messages,
  model = "mistral-tiny",
  dry_run = FALSE,
  ...,
  error_call = current_env()
)

stream(
  messages,
  model = "mistral-tiny",
  dry_run = FALSE,
  ...,
  error_call = current_env()
)
}
\arguments{
\item{messages}{Messages}

\item{model}{which model to use. See \code{\link[=models]{models()}} for more information about which models are available}

\item{dry_run}{if TRUE the request is returned instead of performed}

\item{...}{These dots are for future extensions and must be empty.}

\item{error_call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\value{
A tibble with columns \code{role} and \code{content} with class \code{chat_tibble} or a request
if this is a \code{dry_run}
}
\description{
Chat with the Mistral api

stream
}
\examples{
chat("Top 5 R packages", dry_run = TRUE)

}
