% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{models}
\alias{models}
\title{Retrieve all models available in the Mistral API}
\usage{
models()
}
\value{
A character vector with the models available in the Mistral API
}
\description{
Retrieve all models available in the Mistral API
}
\examples{
models()

}
